// src/app/data-access/models/device.model.ts
export enum BatteryAttribute {
  BatteryStatus = "Battery Status",
  BatteryLevel = "Battery Level",
  Battery = "Battery",
}

export enum BatteryLevel {
  High = "High",
  Medium = "Medium",
  Low = "Low",
  Critical = "Critical",
}

export enum BatteryStatus {
  GoodBattery = "good battery",
  LowBattery = "low battery",
  VeryGoodBattery = "very good",
  GoodBatteryLor = "good",
  LowBatteryLor = "low",
  UltraLowBattery = "ultra low",
}

export enum PowerStatus {
  Off = "off",
  On = "on",
}

export enum WirelessStatus {
  WiFi = "wifi",
  RSSI = "RSSI",
  SIGFOX = "sigfox",
  LORA = "lora",
  NB = "nb-iot",
  SNR = "SNR",
}

export interface IiconData {
  icon: string[];
  color: string;
}

export enum DeviceMode {
  WiFiMac = "wifi 2macs scanned",
  GPSScanned = "gps scanned",
  WiFi3Mac = "3 wifi macs scanned",
  WiFi2Mac = "2 wifi macs scanned",
}

export enum LinkQuality {
  EXCELLENT = "EXCELLENT",                        
  GOOD = "GOOD",
  AVERAGE = "AVERAGE",
  LIMIT = "LIMIT",
}

export enum DoorAttr {
  mode = "Mode",
  lastDorrDur = "LAST_DOOR_OPEN_DURATION",
  alarm = "ALARM",
  battery = "Battery",
  doorOpenTimes = "DOOR_OPEN_TIMES",
  doorStatus = "DOOR_OPEN_STATUS",
}

export enum DeviceAttributeStatus {
  doorStatus = "DOOR_OPEN_STATUS",
  WaterLeakageStatus = "WATER_LEAK_STATUS",
}